Big Swing + Depressure Changes
Increased height of depressure from 70 to 80 (at Gala's request).
Changed big swing so it checks the other blocks it will break with the original block and only breaks them if they match.
Or, that's the intent. Currently it does this after the block breaks, which means it compares the other blocks to air.